refactor profile into seperate component
