+++ frontend/src/components/Profile/Profile.module.scss
@@ -1,0 +1,25 @@
+.profileContainer {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    margin-top: 20px;
+    gap: 2rem;
+}
+
+.textFieldContainer {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+.optionButtons {
+    display: flex;
+    justify-content: flex-end;
+    gap: 1rem;
+}
+
+.avatarContainer {
+    display: flex;
+    gap: 1rem
+}
