+++ frontend/src/scenes/ProfileSettings/ProfileSettingsPage.jsx
@@ -2,12 +2,7 @@ import { React, useState, useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styles from './ProfileSettings.module.scss';
 import classNames from 'classnames';
-import Button from '../../components/Button/Button';
-import CustomTextField from '../../components/TextFieldComponents/CustomTextField/CustomTextField';
-import { InputLabel, Divider } from "@mui/material";
-import Avatar from '../../components/Avatar/Avatar';
-import DeleteConfirmationModal from '../../components/Modals/DeleteConfirmationModal/DeleteConfirmationModal';
-import UploadModal from '../../components/Modals/UploadImageModal/UploadModal';
+import Profile from '../../components/Profile/Profile';
 
 
 const ProfileSettingsPage = ({ title = '', leftContent = () => null, rightContent = () => null, leftAppearance = () => null, onSave = () => null }) => {
@@ -16,13 +11,7 @@ const ProfileSettingsPage = ({ title = '', leftContent = () => null, rightConten
     const handleButtonClick = (index) => {
         setActiveButton(index);
     };
-    const [openDeleteConfirmationModal, setOpenDeleteConfirmationModal] = useState(false);
-    const handleOpenDeleteConfirmationModal = () => setOpenDeleteConfirmationModal(true);
-    const handleCloseDeleteConfirmationModal = () => setOpenDeleteConfirmationModal(false);
-
-    const [openUploadModal, setOpenUploadModal] = useState(false);
-    const handleOpenUploadModal = () => setOpenUploadModal(true);
-    const handleCloseUploadModal = () => setOpenUploadModal(false);
+    
 
     const buttons = ['Profile', 'Password', 'Team'];
 
@@ -44,46 +33,7 @@ const ProfileSettingsPage = ({ title = '', leftContent = () => null, rightConten
                             </button>
                         ))}
                     </div>
-                    {activeButton == 0 &&
-                        (
-                            <div style={{ display: 'flex', flexDirection: 'column', width: '100%', marginTop: '20px', gap: '2rem' }}>
-                                <div style={{ width: '100%', display: 'flex', flexDirection: 'column', gap: '1rem' }}>
-                                    <CustomTextField TextFieldWidth="100%" style={{ display: 'grid', gridTemplateColumns: 'repeat(2, 1fr)', alignItems: 'center' }} labelText='First Name' />
-                                    <CustomTextField TextFieldWidth="100%" style={{ display: 'grid', gridTemplateColumns: 'repeat(2, 1fr)', alignItems: 'center' }} labelText='Last Name' />
-                                    <CustomTextField labelSubText="This is your current email address — it cannot be changed." TextFieldWidth="100%" style={{ display: 'grid', gridTemplateColumns: 'repeat(2, 1fr)', alignItems: 'center' }} labelText='Email' />
-                                    <div style={{ display: 'grid', gridTemplateColumns: 'repeat(2, 1fr)', alignItems: 'center' }}>
-                                        <div>
-                                            <InputLabel sx={{ fontWeight: '600', margin: 0 }}>Your Photo</InputLabel>
-                                            <InputLabel sx={{ fontWeight: '400', fontSize: '13px', margin: 0 }}>This photo will be displayed in your profile page.</InputLabel>
-                                        </div>
-                                        <div style={{ display: 'flex', gap: '1rem' }}>
-                                            <Avatar src='/vendetta.png' />
-                                            <Button buttonType='secondary-grey' variant='contained' text='Delete' onClick={onSave} />
-                                            <Button buttonType='secondary-grey' variant='contained' text='Update' onClick={handleOpenUploadModal} />
-                                        </div>
-                                    </div>
-                                </div>
-                                <div className={styles.optionButtons}>
-                                    <Button text='Save' onClick={onSave} />
-                                </div>
-                                <Divider />
-                                <div>
-                                    <h3>Delete Account</h3>
-                                    <p>Note that deleting your account will remove all data from our system. This is permanent and non-recoverable.</p>
-                                    <Button onClick={handleOpenDeleteConfirmationModal} buttonType='error' variant='contained' text='Delete Account' />
-                                </div>
-                                <UploadModal
-                                    open={openUploadModal}
-                                    handleClose={handleCloseUploadModal}
-                                />
-                                <DeleteConfirmationModal
-                                    open={openDeleteConfirmationModal}
-                                    handleClose={handleCloseDeleteConfirmationModal}
-                                />
-                            </div>
-
-                        )
-                    }
+                    {activeButton == 0 && <Profile onSave={onSave}/>}
                     <div className={styles.leftRightContent}>
                         {activeButton === 1 ? leftAppearance() : leftContent()}
                         {rightContent()}
